+++ inc/storage/global/storage.php
@@ -4,6 +4,7 @@ abstract class GlobalStorage
 {
     abstract public static function path($url);
     abstract public static function copy($url);
+    abstract public static function download($url);
     abstract public static function moveUploadedFile($field);
     abstract public static function getThumbnails($media);
     abstract public static function createThumbnails($media);